Remove mockData() from test-utils.js
