+++ frontend/src/components/__tests__/Test/List.js
@@ -1,9 +1,9 @@
 /* eslint-env jest */
 import React from 'react';
-import { waitForElement } from 'react-testing-library';
-import { renderApollo, mockData } from '../../../utils/test-utils';
+import { waitForElement, render } from 'react-testing-library';
+import { renderApollo } from '../../../utils/test/test-utils';
 import TestList from '../../Test/List';
-import { allTests } from '../../Test/ListConnector';
+import FakeDataProvider from '../../../utils/test/FakeDataProvider';
 
 // <Link /> usa 'next/router' pra definir se o link está ativo (páginal atual)
 jest.mock('next/router');
@@ -15,19 +15,13 @@ describe('<TestList />', () => {
   });
 
   test('renders some tests', async () => {
-    const data = await mockData(allTests);
-    // console.log(JSON.stringify(data, null, 2));
+    const { container, getByLabelText } = render(
+      <FakeDataProvider>
+        <TestList />
+      </FakeDataProvider>,
+    );
 
-    const mocks = [
-      {
-        request: { query: allTests },
-        result: { data },
-      },
-    ];
-
-    const { container, getByText } = renderApollo(<TestList />, { mocks });
-
-    await waitForElement(() => getByText(data.tests[0].title));
-    expect(container.querySelector('ul').children.length).toBe(data.tests.length);
+    await waitForElement(() => getByLabelText('Editar prova'));
+    expect(container.querySelector('ul').children.length).toBe(2);
   });
 });
