+++ frontend/src/utils/test/FakeDataProvider.js
@@ -4,13 +4,46 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 import React from 'react';
 import PropTypes from 'prop-types';
-import { addMockFunctionsToSchema } from 'graphql-tools';
+import { addMockFunctionsToSchema, makeExecutableSchema, MockList } from 'graphql-tools';
 import ApolloClient from 'apollo-client';
 import { SchemaLink } from 'apollo-link-schema';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { ApolloProvider } from 'react-apollo';
-import { globalMocks, schema } from './test-utils';
+import gql from 'graphql-tag';
+import { importSchema } from 'graphql-import';
+import path from 'path';
+import faker from 'faker/locale/pt_BR';
 import mergeResolvers from './mergeResolvers';
+
+// We have to use importSchema() because schema.graphql imports from
+// 'backend/src/__generated__/prisma.graphql'
+const schemaPath = path.join(__dirname, '../../../../backend/src/schema.graphql');
+const typeDefs = gql(importSchema(schemaPath));
+
+// Make a GraphQL schema with no resolvers
+// https://www.apollographql.com/docs/apollo-server/v2/api/graphql-tools.html#makeExecutableSchema
+const schema = makeExecutableSchema({
+  resolverValidationOptions: { requireResolversForResolveType: false },
+  typeDefs,
+});
+
+const globalMocks = {
+  // Makes all ID types random ids instead of a hello world string
+  ID: () => faker.random.uuid(),
+  // Uses 'faker' to generate more meaninful data
+  Test: () => ({
+    title: () => faker.lorem.sentence(),
+    questions: () => new MockList(10),
+  }),
+  Question: () => ({
+    wording: () => faker.lorem.sentences(),
+    imageFullUrl: () => faker.image.technics(300, 300),
+    secondaryWording: () => faker.lorem.sentences(),
+  }),
+  QuestionChoice: () => ({
+    text: () => faker.hacker.phrase(),
+  }),
+};
 
 // globalMocks uses fake data generated by faker.js
 // It is merged with customResolvers, which can be used to customize the data
